README: Add desktop notifications instructions
I also fixed the Heroku dyno scaling command, since it was not working for me at the time.

Sorry @robcole for not doing pull requests initially; I thought commits automatically pushed to the server, but found that they do not. Basically, I wanted to make scheduler alert users, and this was my solution.